Selection improved.

+++ src/pixelinfo.h
@@ -23,7 +23,7 @@ struct sPixelInfo
     cVector<float> point;
     int img_w, img_h;
     unsigned char r, g, b, a;
-    CRect<float> rc;
+    CRect<int> rc;
 };
 
 class CPixelInfo
